// src/shared/api/models/BudgetAccountingSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type BudgetAccountingSchema = {
  /**
   * @type number
   */
  id: number
  /**
   * @type string
   */
  name: string
  /**
   * @type number
   */
  value: number
  /**
   * @type string
   */
  date: string
  /**
   * @type boolean
   */
  planned: boolean
  /**
   * @type string
   */
  variant: string
}